chore: move preferWroomBase to lib
