+++ packages/slice-machine/server/src/api/services/setShortId.ts
@@ -2,7 +2,7 @@ import DefaultClient from "@lib/models/common/http/DefaultClient";
 import { BackendEnvironment } from "@lib/models/common/Environment";
 import { FileSystem } from "@slicemachine/core";
 import { UserProfile } from "@slicemachine/core/build/src/models";
-import preferWroomBase from "../../../../src/utils/preferWroomBase";
+import preferWroomBase from "../../../../lib/utils/preferWroomBase";
 
 export async function setShortId(
   env: BackendEnvironment,
