+++ packages/slice-machine/tests/src/utils/preferWroomBase.test.ts
@@ -1,4 +1,4 @@
-import preferWroomBase from "../../../src/utils/preferWroomBase";
+import preferWroomBase from "../../../lib/utils/preferWroomBase";
 
 describe("src/utils/preferWroomBase", () => {
   test("when endpoint form env.manifest.apiEndpoint is not from wroom.io or wroom.test it should return the base in the from enn.prismicData.base (prismic config)", () => {
